组织机构，客户，客户分类，客户等级，单位，商品分类

+++ kernel/src/main/java/com/flyemu/share/entity/CustomersLevelPrice.java
@@ -34,7 +34,7 @@ public class CustomersLevelPrice implements Serializable {
     private Integer id;
 
     @Comment("商品id")
-    private Integer productsId;
+    private Long productsId;
 
     @Comment("客户分类id")
     private Integer customersLevelId;
